Add function in utils to generate a sample of maximum eigenvalues of an ensemble
