+++ skrmt/ensemble/tests/test_base_ens.py
@@ -62,9 +62,9 @@ def test_base_set_eigval_norm_const():
 
 
 def test_base_ens_plot():
-        '''Testing WignerSemicircleDistribution plot pdf
-        '''
-        fig_name = "test_base_ens_plot_eigval_hist.png"
-        goe = GaussianEnsemble(beta=1, n=100, use_tridiagonal=False)
-        goe.plot_eigval_hist(fig_path=TMP_DIR_PATH+"/"+fig_name)
-        assert os.path.isfile(os.path.join(TMP_DIR_PATH, fig_name)) == True
+    """Testing plot eigval hist
+    """
+    fig_name = "test_base_ens_plot_eigval_hist.png"
+    goe = GaussianEnsemble(beta=1, n=100, use_tridiagonal=False)
+    goe.plot_eigval_hist(fig_path=TMP_DIR_PATH+"/"+fig_name)
+    assert os.path.isfile(os.path.join(TMP_DIR_PATH, fig_name)) == True
